+++ lib/go/edgecontext/edgecontexttest/runner.go
@@ -15,14 +15,6 @@ import (
 )
 
 const (
-	// copied from https://github.com/reddit/edgecontext.py/blob/420e58728ee7085a2f91c5db45df233142b251f9/tests/edge_context_tests.py#L55-L58
-	//headerWithNoAuth                     = "\x0c\x00\x01\x0b\x00\x01\x00\x00\x00\x0bt2_deadbeef\n\x00\x02\x00\x00\x00\x00\x00\x01\x86\xa0\x00\x0c\x00\x02\x0b\x00\x01\x00\x00\x00\x08beefdead\x00\x0c\x00\x04\x0b\x00\x01\x00\x00\x00$becc50f6-ff3d-407a-aa49-fa49531363be\x00\x00"
-	//headerWithValidAuth                  = "\x0c\x00\x01\x0b\x00\x01\x00\x00\x00\x0bt2_deadbeef\n\x00\x02\x00\x00\x00\x00\x00\x01\x86\xa0\x00\x0c\x00\x02\x0b\x00\x01\x00\x00\x00\x08beefdead\x00\x0b\x00\x03\x00\x00\x01\xaeeyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiJ0Ml9leGFtcGxlIiwiZXhwIjoyNTI0NjA4MDAwfQ.dRzzfc9GmzyqfAbl6n_C55JJueraXk9pp3v0UYXw0ic6W_9RVa7aA1zJWm7slX9lbuYldwUtHvqaSsOpjF34uqr0-yMoRDVpIrbkwwJkNuAE8kbXGYFmXf3Ip25wMHtSXn64y2gJN8TtgAAnzjjGs9yzK9BhHILCDZTtmPbsUepxKmWTiEX2BdurUMZzinbcvcKY4Rb_Fl0pwsmBJFs7nmk5PvTyC6qivCd8ZmMc7dwL47mwy_7ouqdqKyUEdLoTEQ_psuy9REw57PRe00XCHaTSTRDCLmy4gAN6J0J056XoRHLfFcNbtzAmqmtJ_D9HGIIXPKq-KaggwK9I4qLX7g\x0c\x00\x04\x0b\x00\x01\x00\x00\x00$becc50f6-ff3d-407a-aa49-fa49531363be\x00\x0c\x00\x05\x0b\x00\x01\x00\x00\x00\tbaseplate\x00\x0c\x00\x06\x0b\x00\x01\x00\x00\x00\x02OK\x00\x0c\x00\x07\x0b\x00\x01\x00\x00\x00\x24" + expectedRequestID + "\x00\f\x00\b\v\x00\x01\x00\x00\x00\x05en_US\x00\x00"
-	//headerWithExpiredAuth                = "\x0c\x00\x01\x0b\x00\x01\x00\x00\x00\x0bt2_deadbeef\n\x00\x02\x00\x00\x00\x00\x00\x01\x86\xa0\x00\x0c\x00\x02\x0b\x00\x01\x00\x00\x00\x08beefdead\x00\x0b\x00\x03\x00\x00\x01\xaeeyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiJ0Ml9leGFtcGxlIiwiZXhwIjoxMjYyMzA0MDAwfQ.iUD0J2blW-HGtH86s66msBXymCRCgyxAZJ6xX2_SXD-kegm-KjOlIemMWFZtsNv9DJI147cNP81_gssewvUnhIHLVvXWCTOROasXbA9Yf2GUsjxoGSB7474ziPOZquAJKo8ikERlhOOVk3r4xZIIYCuc4vGZ7NfqFxjDGKAWj5Tt4VUiWXK1AdxQck24GyNOSXs677vIJnoD8EkgWqNuuwY-iFOAPVcoHmEuzhU_yUeQnY8D-VztJkip5-YPEnuuf-dTSmPbdm9ZTOP8gjTsG0Sdvb9NdLId0nEwawRy8CfFEGQulqHgd1bqTm25U-NyXQi7zroi1GEdykZ3w9fVNQ\x0c\x00\x07\x00\x00"
-	//headerWithAnonAuth                   = "\x0c\x00\x01\x0b\x00\x01\x00\x00\x00\x0bt2_deadbeef\n\x00\x02\x00\x00\x00\x00\x00\x01\x86\xa0\x00\x0c\x00\x02\x0b\x00\x01\x00\x00\x00\x08beefdead\x00\x0b\x00\x03\x00\x00\x01\xc0eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9.eyJyb2xlcyI6WyJhbm9ueW1vdXMiXSwic3ViIjpudWxsLCJleHAiOjI1MjQ2MDgwMDB9.gQDiVzOUh70mKKK-YBTnLHWBOEuQyRllEE1-EIMfy3x5K8PsH9FB6Oy9S5HbILjfGFNrIBeux9HyW6hBDikoZDhn5QWyPNitL1pzMNONGGrXzSfaDoDbFy4MLD03A7zjG3qWBn_wLjgzUXX6qVX6W_gWO7dMqrq0iFvEegue-xQ1HGiXfPgnTrXRRovUO3JHy1LcZsmOjltYj5VGUTWXodBM8ObKEealDxg8yskEPy0IuujNMmb9eIyuHB8Ozzpg-lr790lxP37s5HCf18vrZ-IhRmLcLCqm5WSFyq_Ld2ByblBKL9pPst1AZYZTXNRIqovTAqr6v0-xjUeJ1iho9A\x0c\x00\x07\x00\x00"
-	//headerWithValidServiceAuth           = "\x0c\x00\x01\x00\x0c\x00\x02\x00\x0b\x00\x03\x00\x00\x02\x0aeyJhbGciOiJSUzI1NiIsImtpZCI6IlNIQTI1NjpsWjBoa1dSc0RwYXBlQnUyZWtYOVdZMm9ZSW5Id2RSYVhUd3RCZWNEaWNJIiwidHlwIjoiSldUIn0.eyJzdWIiOiJzZXJ2aWNlL3Rlc3Qtc2VydmljZSIsImV4cCI6MjUyNDYwODAwMH0.P41Iahxu-Bbg5srTFSQTBkzwiff4ytlhVBUYuyYTFGY_7XCyKdZywUmVHRY_Q2w8Q2uaybnmuoM95JhRpdNYcTPIYWEby4Z5DSV-zMqqmHnP22aH_sAckFQl86Yw_2pdZpKKJ-KQkyT0vEkxe-vNs5HhEdBr6Rae0g2SKEr7RaPMoToq6xpucDAREVWa7yJMtyyNtiVixeLoxTegRLOZTFEVt4TTYKDuT2FdEY5P2b8BOSpFMoiv9w51gZO1qvn9Zjrl00Z-lI_onihMIkrG_viWVAlzEl8d5ZWuJVjHJvm7O0CS4OuhZocE2qbYQrw9THSS1Mh4YR-_r2v1ArYnVA\x0c\x00\x04\x00\x0c\x00\x05\x0b\x00\x01\x00\x00\x00\x0eorigin/service\x00\x0c\x00\x06\x00\x0c\x00\x07\x0b\x00\x01\x00\x00\x00$1566dce9-9567-4952-b23b-9fd72e111162\x00\x00"
-	//headerWithValidServiceAuthAndOptions = "\x0c\x00\x01\x00\x0c\x00\x02\x00\x0b\x00\x03\x00\x00\x02VeyJhbGciOiJSUzI1NiIsImtpZCI6IlNIQTI1NjpsWjBoa1dSc0RwYXBlQnUyZWtYOVdZMm9ZSW5Id2RSYVhUd3RCZWNEaWNJIiwidHlwIjoiSldUIn0.eyJzdWIiOiJzZXJ2aWNlL3Rlc3Qtc2VydmljZSIsImV4cCI6MjUyNDYwODAwMCwib2JvIjp7ImFpZCI6InQyX2RlYWRiZWVmIiwicm9sZXMiOlsiYWRtaW4iXX0sInNlYSI6dHJ1ZX0.PVefAKWUFfk_7QKen6Iz0Cfu95Yp92lYETlrxCUacLsa9u-qz36aet21iwFrdnJiz7gDeJRH7sOJyh6jRmkD0ptWs4Zl7VqpZY-ALgDOdhwSHoUIoV2L7twT-Dm3Tdyfbzq01fOni9ioq5akKnETC5IbLSOqp1ssWJcgo_9g-X-SdRiuf5u8YHD2Mrep5U21bkbYnm4rK9tX_oCnhrrp4rbXi5yogx594oNmOWUedIeyv6QY_xVGbaXOz7deBIWQY2fSYG3cpiBNtSYEJ4yDTbjGY0G1Vp78bX8YZlboc13TGoDpARdfHuHeQU0wAQEhi7pu0Q4FufEVua4q1f0P3A\x0c\x00\x04\x00\x0c\x00\x05\x0b\x00\x01\x00\x00\x00\x0eorigin/service\x00\x0c\x00\x06\x00\x0c\x00\x07\x0b\x00\x01\x00\x00\x00$a3b2d5c2-ab27-4948-9dae-78a3ffb46957\x00\x00"
-
 	headerWithNoAuth                     = "\f\x00\x01\v\x00\x01\x00\x00\x00\vt2_deadbeef\n\x00\x02\x00\x00\x00\x00\x00\x01\x86\xa0\x00\f\x00\x02\v\x00\x01\x00\x00\x00\bbeefdead\x00\v\x00\x03\x00\x00\x02\x11eyJhbGciOiJSUzI1NiIsImtpZCI6IlNIQTI1NjphaGxBYkE5SW9qM1VGTUhzOEc1ZWFia1JBSG1hWlpaTm9OREJZQW40OHBVIiwidHlwIjoiSldUIn0.eyJleHAiOjE3NDc4ODkzODksIm5iZiI6MTc0Nzg4OTI2OSwiaWF0IjoxNzQ3ODg5MzI5fQ.uRnz-pWFp6Fj3vgGnZPDds9pBcWdh2G3vJm1jLMKTEDswlVS0leZ0XSvGzN5JbX9Em7NVL3-Gb0NMtvDkC7TOcIlHWxLD891ooOtDDJwIubUuHnIQr5h4x6DhdmLxu0wYB38v_JrJVHlLywjK_YdXVA-Bw6mgq9inxB6PzB4BrH3Q2upU4IpGsrPNjMk6hD2qUaAwBi7xKWgsu8D3xHxTXQ1VZBAnC3QGEyE1WYABNDzNAYzsmZ5-M6ke3syAAUtAbweUWf6tUQ2TuTY_Nxnl87q1V9a4y27ecB6NVJdzrCDw8yrF2cSSEEWX6Yi6FDxdHJ91bwlXBbZIwAwMI5SZA\f\x00\x04\v\x00\x01\x00\x00\x00$becc50f6-ff3d-407a-aa49-fa49531363be\x00\f\x00\x05\x00\f\x00\x06\x00\x00"
 	headerWithExpiredAuth                = "\f\x00\x01\v\x00\x01\x00\x00\x00\vt2_deadbeef\n\x00\x02\x00\x00\x00\x00\x00\x01\x86\xa0\x00\f\x00\x02\v\x00\x01\x00\x00\x00\bbeefdead\x00\v\x00\x03\x00\x00\x02\x11eyJhbGciOiJSUzI1NiIsImtpZCI6IlNIQTI1NjphaGxBYkE5SW9qM1VGTUhzOEc1ZWFia1JBSG1hWlpaTm9OREJZQW40OHBVIiwidHlwIjoiSldUIn0.eyJleHAiOjE3NDc4ODkzODksIm5iZiI6MTc0Nzg4OTI2OSwiaWF0IjoxNzQ3ODg5MzI5fQ.uRnz-pWFp6Fj3vgGnZPDds9pBcWdh2G3vJm1jLMKTEDswlVS0leZ0XSvGzN5JbX9Em7NVL3-Gb0NMtvDkC7TOcIlHWxLD891ooOtDDJwIubUuHnIQr5h4x6DhdmLxu0wYB38v_JrJVHlLywjK_YdXVA-Bw6mgq9inxB6PzB4BrH3Q2upU4IpGsrPNjMk6hD2qUaAwBi7xKWgsu8D3xHxTXQ1VZBAnC3QGEyE1WYABNDzNAYzsmZ5-M6ke3syAAUtAbweUWf6tUQ2TuTY_Nxnl87q1V9a4y27ecB6NVJdzrCDw8yrF2cSSEEWX6Yi6FDxdHJ91bwlXBbZIwAwMI5SZA\f\x00\x04\x00\f\x00\x05\x00\f\x00\x06\x00\x00"
 	headerWithAnonAuth                   = "\f\x00\x01\v\x00\x01\x00\x00\x00\vt2_deadbeef\n\x00\x02\x00\x00\x00\x00\x00\x01\x86\xa0\x00\f\x00\x02\v\x00\x01\x00\x00\x00\bbeefdead\x00\v\x00\x03\x00\x00\x02.eyJhbGciOiJSUzI1NiIsImtpZCI6IlNIQTI1NjphaGxBYkE5SW9qM1VGTUhzOEc1ZWFia1JBSG1hWlpaTm9OREJZQW40OHBVIiwidHlwIjoiSldUIn0.eyJleHAiOjI1MjQ2MDgwMDAsIm5iZiI6MTc0Nzg4OTI2OSwiaWF0IjoxNzQ3ODg5MzI5LCJyb2xlcyI6WyJhbm9ueW1vdXMiXX0.TmeI4HaJmwfVziCMmodL7M6h_UbBYSv_0_cFjRtq8F9YAxjaOY0s_fXGuuBx1SkFzKpxyV8nAowHQ82xuPRq0X4El4E86otGi87k9yM9Ij5_c5Xjud8_XuH2Hmb8GIMJ35MaPPjtKMSpB9r-kxUktB-P55LhI8z_FfGpKd3lCEhyz1GLSnkLiIP7l72Ertz722YWOpN5jR2R2dftH8_4Om8PC7WnJNG2iGZL6V_9g6csy3e_96C1iiFGx2WSNhdGHIDC1OULaobHNa4JFdwf7IglXLgUboekunyx-4s8pVaL36oNMbGQtAhuwZ_rM0gMV3Hy1wgg_ZOMadBQiDi2Lw\f\x00\x04\x00\f\x00\x05\x00\f\x00\x06\x00\x00"
@@ -43,9 +35,6 @@ const (
 		"\x00")
 
 	validToken = "eyJhbGciOiJSUzI1NiIsImtpZCI6IlNIQTI1NjphaGxBYkE5SW9qM1VGTUhzOEc1ZWFia1JBSG1hWlpaTm9OREJZQW40OHBVIiwidHlwIjoiSldUIn0.eyJzdWIiOiJ0Ml9leGFtcGxlIiwiZXhwIjoyNTI0NjA4MDAwLCJuYmYiOjE3NDc4ODkyNjksImlhdCI6MTc0Nzg4OTMyOX0.sxfTZVoT-IBLlHQQ7frsCXlXyBnqqVueR7ZKh8KjrkDMzUMGSw4xjIXmpKLCASj8OVoVQ8una3kQASM18gcmLcwZXoJEiaYBWyB7TYu-OXL2prX9zJbBWVoPge_iUZ86Qt2ddJE5LxTDpVZ0BkdmDe9jE2j2RE5igyLaydfLdtkbNHDuvxeHnKFhj51qeez5Y3kf6fmJk-dAItKRFunQ_mCDPzvMsAwq884bw2ayh_M9YH1M7KfuYPi8x-vJwMgqVTOIXgLYoPWkQ2ZPfaLmrqfGHzdKoD6dfuOG5jnbZzkXOg3QZ5uPyRMNERkarfmynDaR-YePReFbLpE-1Kpwng"
-
-	// copied from https://github.com/reddit/edgecontext.py/blob/420e58728ee7085a2f91c5db45df233142b251f9/tests/edge_context_tests.py#L54
-	//validToken = `eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiJ0Ml9leGFtcGxlIiwiZXhwIjoyNTI0NjA4MDAwfQ.dRzzfc9GmzyqfAbl6n_C55JJueraXk9pp3v0UYXw0ic6W_9RVa7aA1zJWm7slX9lbuYldwUtHvqaSsOpjF34uqr0-yMoRDVpIrbkwwJkNuAE8kbXGYFmXf3Ip25wMHtSXn64y2gJN8TtgAAnzjjGs9yzK9BhHILCDZTtmPbsUepxKmWTiEX2BdurUMZzinbcvcKY4Rb_Fl0pwsmBJFs7nmk5PvTyC6qivCd8ZmMc7dwL47mwy_7ouqdqKyUEdLoTEQ_psuy9REw57PRe00XCHaTSTRDCLmy4gAN6J0J056XoRHLfFcNbtzAmqmtJ_D9HGIIXPKq-KaggwK9I4qLX7g`
 
 	verificationSecret = "-----BEGIN PUBLIC KEY-----\nMIIBIjANBgkqhkiG9w0BAQEFAAOCAQ8AMIIBCgKCAQEAziSyW/ychrelzKu0unCI\npuo3eXokDiywKoUSKlvWOWujJZHWWakIxWuMWZu+L756dVYcgjJfQM/pZo3AlPxy\nuQICiuKH9Ol1tFJV+bdd0IC2uJKSaqZLXslK+cQpPQdnWTmBTR1TXsqSQx3kuRfq\nQjF4R6Yr7mdSIX8/uZg82WIVCR24++G8/ENSNJ+H2Gi/YujO/3Qc5ujPwpwIsTkg\nPQLxotvYbQsRyUZfuA6EQlUZpKbJoBUriNb0+Z+QANTD/hJjUNws/NNSHI6aWfhW\ngMQQSTmd9L5pxyY2m5kzJgMXQKWHlj4PjaKo5TqS+SmN2Vm753oUNFGsBybDNPAv\n2wIDAQAB\n-----END PUBLIC KEY-----\n"
 )
@@ -254,24 +243,29 @@ type runnerConfig struct {
 	backend           *edgecontext.Backend
 }
 
+// RunnerOption is a function that configures the test Runner.
 type RunnerOption func(*runnerConfig)
 
-func WithBackend(t *testing.T, backend *edgecontext.Backend) RunnerOption {
+// WithBackend sets the backend for the Runner.
+func WithBackend(backend *edgecontext.Backend) RunnerOption {
 	return func(c *runnerConfig) {
 		c.backend = backend
 	}
 }
 
+// WithDefaultBackend configures the Runner to use the default backend.
 func WithDefaultBackend() RunnerOption {
 	return func(c *runnerConfig) {
 		c.useDefaultBackend = true
 	}
 }
 
+// Runner is a test runner for edgecontext package. This can be used to run tests to verify that a new Backend works.
 type Runner struct {
 	impl *edgecontext.Impl
 }
 
+// NewRunner creates a new Runner instance for testing. It requires either a backend to be set or the use of WithDefaultBackend.
 func NewRunner(t *testing.T, opts ...RunnerOption) *Runner {
 	var cfg runnerConfig
 	for _, opt := range opts {
@@ -307,6 +301,7 @@ func NewRunner(t *testing.T, opts ...RunnerOption) *Runner {
 	}
 }
 
+// Run runs all the tests defined in the Runner.
 func (r *Runner) Run(t *testing.T) {
 	for name, fn := range map[string]func(*testing.T){
 		"TestNew":        r.TestNew,
@@ -317,6 +312,7 @@ func (r *Runner) Run(t *testing.T) {
 	}
 }
 
+// TestNew tests the creation of a new EdgeContext with valid parameters.
 func (r *Runner) TestNew(t *testing.T) {
 	ctx := context.Background()
 	e, err := edgecontext.New(
@@ -354,6 +350,7 @@ func (r *Runner) TestNew(t *testing.T) {
 	}
 }
 
+// TestLocale tests the validation of locale codes in EdgeContext.
 func (r *Runner) TestLocale(t *testing.T) {
 	for _, c := range []struct {
 		label  string
@@ -418,6 +415,7 @@ func (r *Runner) TestLocale(t *testing.T) {
 	}
 }
 
+// TestFromHeader tests the creation of an EdgeContext from a header string.
 func (r *Runner) TestFromHeader(t *testing.T) {
 
 	t.Run(
