+++ lib/go/edgecontext/edgecontexttest/transformer.go
@@ -45,76 +45,74 @@ type edgeContext struct {
 	Service       service
 }
 
+// Diff compares two EdgeRequestContext objects and returns a string if any differences are found.
 func Diff(want, got *edgecontext.EdgeRequestContext) string {
 	return cmp.Diff(
 		want, got,
-		Transform(),
+		transform(),
 		cmpopts.EquateEmpty(),
 		cmpopts.SortSlices(func(a, b string) bool { return a < b }),
 	)
 }
 
-func Transform() cmp.Option {
+func transform() cmp.Option {
 	return cmp.Transformer("EdgeRequestContext", func(ec *edgecontext.EdgeRequestContext) edgeContext {
-		return transform(ec)
-	})
-}
-func transform(ec *edgecontext.EdgeRequestContext) edgeContext {
-	transformed := edgeContext{
-		HasAuthenticationToken: ec.AuthToken() != nil,
-		RequestID:              ec.RequestID(),
-		SessionID:              ec.SessionID(),
-		DeviceID:               ec.DeviceID(),
-		LocaleCode:             ec.LocaleCode(),
-		CountryCode:            ec.CountryCode(),
+		transformed := edgeContext{
+			HasAuthenticationToken: ec.AuthToken() != nil,
+			RequestID:              ec.RequestID(),
+			SessionID:              ec.SessionID(),
+			DeviceID:               ec.DeviceID(),
+			LocaleCode:             ec.LocaleCode(),
+			CountryCode:            ec.CountryCode(),
 
-		OriginService: originService{
-			Name: ec.OriginService().Name(),
-		},
-	}
+			OriginService: originService{
+				Name: ec.OriginService().Name(),
+			},
+		}
 
-	if id, ok := ec.User().ID(); ok {
-		loid, _ := ec.User().LoID()
-		createdAt, _ := ec.User().CookieCreatedAt()
-		transformed.User = user{
-			ID:         id,
-			IsLoggedIn: ec.User().IsLoggedIn(),
-			LoID:       loid,
-			CreatedAt:  createdAt,
-			Roles:      ec.User().Roles(),
-		}
-	} else {
-		loid, _ := ec.User().LoID()
-		createdAt, _ := ec.User().CookieCreatedAt()
-		transformed.User = user{
-			ID:         "",
-			IsLoggedIn: ec.User().IsLoggedIn(),
-			LoID:       loid,
-			CreatedAt:  createdAt,
-			Roles:      ec.User().Roles(),
-		}
-	}
-
-	if client, ok := ec.OAuthClient(); ok {
-		transformed.OAuthClient = oauthClient{
-			ID:   client.ID(),
-			Type: ec.AuthToken().OAuthClientType,
-		}
-	}
-
-	if svc, ok := ec.Service(); ok {
-		name, _ := svc.Name()
-		transformed.Service = service{
-			Name:                   name,
-			RequestsElevatedAccess: svc.RequestsElevatedAccess(),
-		}
-		if onBehalfOf, ok := svc.OnBehalfOfID(); ok {
-			roles, _ := svc.OnBehalfOfRoles()
-			transformed.Service.OnBehalfOf = user{
-				ID:    onBehalfOf,
-				Roles: roles,
+		if id, ok := ec.User().ID(); ok {
+			loid, _ := ec.User().LoID()
+			createdAt, _ := ec.User().CookieCreatedAt()
+			transformed.User = user{
+				ID:         id,
+				IsLoggedIn: ec.User().IsLoggedIn(),
+				LoID:       loid,
+				CreatedAt:  createdAt,
+				Roles:      ec.User().Roles(),
+			}
+		} else {
+			loid, _ := ec.User().LoID()
+			createdAt, _ := ec.User().CookieCreatedAt()
+			transformed.User = user{
+				ID:         "",
+				IsLoggedIn: ec.User().IsLoggedIn(),
+				LoID:       loid,
+				CreatedAt:  createdAt,
+				Roles:      ec.User().Roles(),
 			}
 		}
-	}
-	return transformed
+
+		if client, ok := ec.OAuthClient(); ok {
+			transformed.OAuthClient = oauthClient{
+				ID:   client.ID(),
+				Type: ec.AuthToken().OAuthClientType,
+			}
+		}
+
+		if svc, ok := ec.Service(); ok {
+			name, _ := svc.Name()
+			transformed.Service = service{
+				Name:                   name,
+				RequestsElevatedAccess: svc.RequestsElevatedAccess(),
+			}
+			if onBehalfOf, ok := svc.OnBehalfOfID(); ok {
+				roles, _ := svc.OnBehalfOfRoles()
+				transformed.Service.OnBehalfOf = user{
+					ID:    onBehalfOf,
+					Roles: roles,
+				}
+			}
+		}
+		return transformed
+	})
 }
