+++ lib/go/edgecontext/edgecontexttest/want.go
@@ -5,16 +5,17 @@ import (
 	"time"
 
 	"github.com/reddit/baseplate.go/timebp"
-	"github.com/reddit/edgecontext/lib/go/ecdata"
 	"github.com/reddit/edgecontext/lib/go/edgecontext"
 )
 
 type authTokenConfig struct {
-	token *ecdata.AuthenticationToken
+	token *edgecontext.AuthenticationToken
 }
 
+// AuthenticationTokenOption is a function that configures the WithAuthenticationToken option for Want.
 type AuthenticationTokenOption func(*authTokenConfig)
 
+// WithLoggedInUser sets the AuthenticationToken for a logged-in user.
 func WithLoggedInUser(id string, createdAt time.Time, roles []string) AuthenticationTokenOption {
 	return func(cfg *authTokenConfig) {
 		cfg.token.RegisteredClaims.Subject = id
@@ -29,6 +30,7 @@ func WithLoggedInUser(id string, createdAt time.Time, roles []string) Authentica
 	}
 }
 
+// WithLoggedOutUser sets the AuthenticationToken for a logged-out user.
 func WithLoggedOutUser(loid string, createdAt time.Time, roles []string) AuthenticationTokenOption {
 	return func(cfg *authTokenConfig) {
 		cfg.token.RegisteredClaims.Subject = ""
@@ -47,14 +49,17 @@ type serviceConfig struct {
 	*authTokenConfig
 }
 
+// WithServiceOption is a function that configures the WithService option for WithAuthenticationToken.
 type WithServiceOption func(*serviceConfig)
 
+// WithServiceRequestsElevatedAccess sets the AuthenticationToken to indicate that the service requests elevated access.
 func WithServiceRequestsElevatedAccess() WithServiceOption {
 	return func(cfg *serviceConfig) {
 		cfg.token.ServiceRequestedElevatedAccess = true
 	}
 }
 
+// WithOnBehalfOfUser sets the AuthenticationToken to indicate that the service is acting on behalf of a user.
 func WithOnBehalfOfUser(id string, roles []string) WithServiceOption {
 	return func(cfg *serviceConfig) {
 		cfg.token.OnBehalfOf = &struct {
@@ -67,6 +72,7 @@ func WithOnBehalfOfUser(id string, roles []string) WithServiceOption {
 	}
 }
 
+// WithService sets the AuthenticationToken for a service.
 func WithService(name string, opts ...WithServiceOption) AuthenticationTokenOption {
 	return func(cfg *authTokenConfig) {
 		cfg.token.RegisteredClaims.Subject = "service/" + name
@@ -78,16 +84,18 @@ func WithService(name string, opts ...WithServiceOption) AuthenticationTokenOpti
 	}
 }
 
+// WithNoAuthenticationToken sets the AuthenticationToken to nil.
 func WithNoAuthenticationToken() WantOption {
 	return func(cfg *wantConfig) {
 		cfg.source.token = nil
 	}
 }
 
+// WithAuthenticationToken sets the AuthenticationToken in the Want configuration.
 func WithAuthenticationToken(opts ...AuthenticationTokenOption) WantOption {
 	return func(cfg *wantConfig) {
 		authCfg := authTokenConfig{
-			token: &ecdata.AuthenticationToken{},
+			token: &edgecontext.AuthenticationToken{},
 		}
 		for _, opt := range opts {
 			opt(&authCfg)
@@ -97,11 +105,13 @@ func WithAuthenticationToken(opts ...AuthenticationTokenOption) WantOption {
 }
 
 type wantConfig struct {
-	source *StubDataSource
+	source *StubHeaderUnmarshaler
 }
 
+// WantOption is a function that configures the Want function.
 type WantOption func(*wantConfig)
 
+// WithLocation sets the locale and country codes in the Want configuration.
 func WithLocation(localeCode, countryCode string) WantOption {
 	return func(cfg *wantConfig) {
 		cfg.source.localeCode = localeCode
@@ -109,12 +119,14 @@ func WithLocation(localeCode, countryCode string) WantOption {
 	}
 }
 
+// WithOriginService sets the origin service name in the Want configuration.
 func WithOriginService(name string) WantOption {
 	return func(cfg *wantConfig) {
 		cfg.source.originService = name
 	}
 }
 
+// WithRequestMetadata sets the request metadata in the Want configuration.
 func WithRequestMetadata(requestID, sessionID, deviceID string) WantOption {
 	return func(cfg *wantConfig) {
 		cfg.source.requestID = requestID
@@ -123,6 +135,7 @@ func WithRequestMetadata(requestID, sessionID, deviceID string) WantOption {
 	}
 }
 
+// WithLoIDCookie sets the non-authentication token LoID and its creation time in the Want configuration.
 func WithLoIDCookie(loid string, createdAt time.Time) WantOption {
 	return func(cfg *wantConfig) {
 		cfg.source.loid = loid
@@ -130,6 +143,7 @@ func WithLoIDCookie(loid string, createdAt time.Time) WantOption {
 	}
 }
 
+// WithOAuthClient sets the OAuth client ID and type in the Want configuration.
 func WithOAuthClient(id, clientType string) WantOption {
 	return func(cfg *wantConfig) {
 		cfg.source.token.OAuthClientID = id
@@ -137,8 +151,9 @@ func WithOAuthClient(id, clientType string) WantOption {
 	}
 }
 
+// Want creates a new edgecontext.EdgeRequestContext to compare against in tests.
 func Want(opts ...WantOption) *edgecontext.EdgeRequestContext {
-	cfg := wantConfig{source: &StubDataSource{}}
+	cfg := wantConfig{source: &StubHeaderUnmarshaler{}}
 	for _, opt := range opts {
 		opt(&cfg)
 	}
