+++ lib/go/edgecontext/validator_middleware_test.go
@@ -1,4 +1,4 @@
-package v0
+package edgecontext
 
 import (
 	"context"
