--- lib/go/internal/v0/validator.go
@@ -1,172 +1,0 @@
-package v0
-
-import (
-	"context"
-	"crypto/rsa"
-	"errors"
-	"fmt"
-	"log/slog"
-
-	"github.com/golang-jwt/jwt/v5"
-	"github.com/reddit/baseplate.go/secrets"
-	"github.com/reddit/edgecontext/lib/go/ecdata"
-	"golang.org/x/crypto/ssh"
-)
-
-type keysType struct {
-	// map of kid -> pub key.
-	m map[string]*rsa.PublicKey
-
-	// when either kid header does not exist in the jwt token,
-	// or the kid is not present in the map,
-	// we fallback to the first (usually current) key.
-	first *rsa.PublicKey
-}
-
-func (kt *keysType) getKey(kid string) *rsa.PublicKey {
-	if key := kt.m[kid]; key != nil {
-		return key
-	}
-	return kt.first
-}
-
-const (
-	authenticationPubKeySecretPath = "secret/authentication/public-key"
-	jwtAlg                         = "RS256"
-)
-
-// JWTHeaderKeyID is the JWT header for the key id,
-// as defined in RFC 7517 section 4.5.
-const JWTHeaderKeyID = "kid"
-
-var (
-	// ErrEmptyToken is an error returned by ValidateToken indicates that the JWT
-	// token is empty string.
-	ErrEmptyToken = errors.New("edgecontext.ValidateToken: empty JWT token")
-
-	// ErrInvalidToken is an error returned by ValidateToken indicates that
-	// the token returned by parsing the JWT was invalid, although an error was
-	// not returned by parsing.
-	ErrInvalidToken = errors.New("edgecontext.ValidateToken: invalid token")
-
-	// ErrInvalidTokenType is an error returned by ValidateToken indicates that
-	// the claims on the token return by parsing the JWT is not a
-	// *AuthenticationToken.
-	ErrInvalidTokenType = errors.New("edgecontext.ValidateToken: invalid token type")
-
-	// ErrNoPublicKeysLoaded is an error returned by ValidateToken indicates that
-	// the function is called before any public keys are loaded from secrets.
-	ErrNoPublicKeysLoaded = errors.New("edgecontext.ValidateToken: no public keys loaded")
-)
-
-// ValidateToken parses and validates a jwt token, and return the decoded
-// AuthenticationToken.
-func (impl *Impl) ValidateToken(token string) (*ecdata.AuthenticationToken, error) {
-	keys, ok := impl.keysValue.Load().(*keysType)
-	if !ok {
-		// This would only happen when all previous middleware parsing failed.
-		return nil, ErrNoPublicKeysLoaded
-	}
-
-	if token == "" {
-		// If we don't do the special handling here,
-		// jwt.ParseWithClaims below will return an error with message
-		// "token contains an invalid number of segments".
-		// Also that's still true, it's less obvious what's actually going on.
-		// Returning different error for empty token can also help highlighting
-		// other invalid tokens that actually causes that invalid number of segments
-		// error.
-		return nil, ErrEmptyToken
-	}
-
-	tok, err := jwt.ParseWithClaims(
-		token,
-		&ecdata.AuthenticationToken{},
-		func(jt *jwt.Token) (interface{}, error) {
-			kid, _ := jt.Header[JWTHeaderKeyID].(string)
-			return keys.getKey(kid), nil
-		},
-		jwt.WithValidMethods([]string{jwtAlg}),
-	)
-	if err != nil {
-		return nil, err
-	}
-
-	if !tok.Valid {
-		return nil, ErrInvalidToken
-	}
-
-	if claims, ok := tok.Claims.(*ecdata.AuthenticationToken); ok {
-		return claims, nil
-	}
-
-	return nil, fmt.Errorf("%w: %T", ErrInvalidTokenType, tok.Claims)
-}
-
-func (impl *Impl) validatorMiddleware(next secrets.SecretHandlerFunc) secrets.SecretHandlerFunc {
-	return func(sec *secrets.Secrets) {
-		defer next(sec)
-
-		versioned, err := sec.GetVersionedSecret(authenticationPubKeySecretPath)
-		if err != nil {
-			slog.ErrorContext(
-				context.Background(),
-				fmt.Sprintf(
-					"Failed to get secrets %q",
-					authenticationPubKeySecretPath,
-				),
-				"err", err,
-			)
-			return
-		}
-
-		keys := parseVersionedKeys(context.Background(), versioned)
-		if keys != nil {
-			impl.keysValue.Store(keys)
-		}
-	}
-}
-
-func parseVersionedKeys(ctx context.Context, versioned secrets.VersionedSecret) *keysType {
-	all := versioned.GetAll()
-	keys := &keysType{
-		m: make(map[string]*rsa.PublicKey, len(all)),
-	}
-	for i, v := range all {
-		key, err := jwt.ParseRSAPublicKeyFromPEM([]byte(v))
-		if err != nil {
-			slog.ErrorContext(
-				ctx, fmt.Sprintf("Failed to parse key #%d", i),
-				"err", err,
-			)
-		} else {
-			if keys.first == nil {
-				keys.first = key
-			}
-			if fingerprint, err := RSAPublicKeyFingerprint(key); err != nil {
-				slog.ErrorContext(
-					ctx, fmt.Sprintf("Failed to get fingerprint of key #%d", i),
-					"err", err,
-				)
-			} else {
-				keys.m[fingerprint] = key
-			}
-		}
-	}
-	if keys.first == nil {
-		slog.InfoContext(ctx, "No valid keys in secrets store.")
-		return nil
-	}
-	return keys
-}
-
-// RSAPublicKeyFingerprint calculates the fingerprint of an RSA public key,
-// using ssh.FingerprintSHA256:
-// https://pkg.go.dev/golang.org/x/crypto/ssh#FingerprintSHA256
-func RSAPublicKeyFingerprint(pubKey *rsa.PublicKey) (string, error) {
-	key, err := ssh.NewPublicKey(pubKey)
-	if err != nil {
-		return "", err
-	}
-	return ssh.FingerprintSHA256(key), nil
-}
